+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## In Git (not released)
 
 * reportportal/reportportal#293 - Re-licence client side to Apache 2.0
+* Remove multipart monkeypatch of RestClient. Replace RestClient with HTTP gem as it supports multipart fine